backend/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
pillow==10.4.0
platformdirs==4.2.2
PyJWT==2.9.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.26.3
